refactor: trip dto -> request, response 디렉터리 구분
