+++ src/main/java/com/pyro/yolog/domain/trip/entity/Trip.java
@@ -1,7 +1,7 @@
 package com.pyro.yolog.domain.trip.entity;
 
 import com.pyro.yolog.domain.member.entity.Member;
-import com.pyro.yolog.domain.trip.dto.TripRequest;
+import com.pyro.yolog.domain.trip.dto.request.TripRequest;
 import com.pyro.yolog.global.config.BaseTimeEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
